# server/src/main/resources/openapi/storyblock-v1.yaml
openapi: 3.1.0
info:
  title: StoryBlock Engine API
  version: 1.1.0
  description: Immutable, evidence-bound narrative editing contracts.
servers:
  - url: /v1
security:
  - bearerAuth: []
...
paths:
  /admin/novels:
    get:
      operationId: listAdminNovels
      summary: List persisted novels for the read-only admin library
      x-required-role: operator
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - name: q
          in: query
          required: false
          schema:
            type: string
            maxLength: 200
            default: ""
      responses:
        "200":
          description: Page of current persisted novel heads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminNovelCatalog"
        default:
          $ref: "#/components/responses/Problem"
  /admin/novels/{novelId}:
    get:
      operationId: readAdminNovel
      summary: Read a persisted novel and its current canonical revision
      x-required-role: operator
      parameters:
        - $ref: "#/components/parameters/NovelId"
      responses:
        "200":
          description: Current persisted novel and complete readable revision
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminNovelDetail"
        default:
          $ref: "#/components/responses/Problem"
...
  /novels/{novelId}:
    get:
      operationId: getNovelHead
      summary: Get current novel head metadata
      x-required-scopes: ["novel:read"]
      parameters:
        - $ref: "#/components/parameters/NovelId"
      responses:
        "200":
          description: Current head metadata
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NovelHead"
        default:
          $ref: "#/components/responses/Problem"
...
  /novels/{novelId}/access-keys:
    post:
      operationId: createAccessKey
      summary: Issue a scoped per-novel bearer credential
      x-required-scopes: ["novel:admin"]
      parameters:
        - $ref: "#/components/parameters/NovelId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessKeyRequest"
      responses:
        "201":
          description: Credential issued; secret is returned only once
          headers:
            Cache-Control:
              description: Prevent storage of the one-time credential response.
              schema:
                const: no-store
            Pragma:
              description: Legacy cache prevention for the one-time credential response.
              schema:
                const: no-cache
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessKeyCreated"
        default:
          $ref: "#/components/responses/Problem"
  /access-keys/{keyId}:
    delete:
      operationId: revokeAccessKey
      summary: Revoke a scoped bearer credential
      x-required-scopes: ["novel:admin"]
      parameters:
        - $ref: "#/components/parameters/KeyId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Key is revoked or was already revoked idempotently
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevokeResponse"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: nv_<public-key-id>.<256-bit-secret>
  parameters:
    NovelId:
      name: novelId
      in: path
      required: true
      schema:
        type: string
        pattern: '^nov_[0-9a-f-]{36}$'
...
    KeyId:
      name: keyId
      in: path
      required: true
      schema:
        type: string
        pattern: '^key_[0-9a-f-]{36}$'
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: A single quoted strong canonical SHA-256 ETag.
      schema:
        type: string
        pattern: '^"sha256:[0-9a-f]{64}"$'
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 200
  headers:
    StrongEtag:
      description: Quoted strong ETag of the canonical revision or immutable resource view.
      required: true
      schema:
        type: string
        pattern: '^"sha256:[0-9a-f]{64}"$'
...
  responses:
    Problem:
      description: RFC problem response with a stable machine code
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, code, detail, instance, request_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]{1,63}$'
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
        current_revision_id:
          type: string
        current_etag:
          type: string
      additionalProperties: true
...
    AdminNovelSummary:
      type: object
      required: [agent_write_registered, block_count, chapter_count, han_character_count, han_text_sha256, head_hash, head_revision_id, head_sequence, language, main_characters, novel_id, scene_count, title, tnt_cannon_count, updated_at, zombie_count]
      properties:
        agent_write_registered:
          type: boolean
        block_count:
          type: integer
          minimum: 0
        chapter_count:
          type: integer
          minimum: 0
        han_character_count:
          type: integer
          minimum: 0
        han_text_sha256:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        head_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        head_revision_id:
          type: string
          pattern: '^rev_[0-9a-f-]{36}$'
        head_sequence:
          type: integer
          minimum: 0
        language:
          type: string
        main_characters:
          type: array
          maxItems: 5
          items:
            type: string
        novel_id:
          type: string
          pattern: '^nov_[0-9a-f-]{36}$'
        scene_count:
          type: integer
          minimum: 0
        title:
          type: string
        tnt_cannon_count:
          type: integer
          minimum: 0
        updated_at:
          type: string
          format: date-time
        zombie_count:
          type: integer
          minimum: 0
      additionalProperties: false
    AdminNovelCatalog:
      type: object
      required: [items, page, schema_version, size, total, total_pages]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminNovelSummary"
        page:
          type: integer
          minimum: 0
        schema_version:
          const: admin-novel-reader-1.0.0
        size:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
      additionalProperties: false
    AdminNovelDetail:
      type: object
      required: [novel, revision, schema_version]
      properties:
        novel:
          $ref: "#/components/schemas/AdminNovelSummary"
        revision:
          $ref: "#/components/schemas/CanonicalRevision"
        schema_version:
          const: admin-novel-reader-1.0.0
      additionalProperties: false
...
    NovelHead:
      type: object
      required: [novel_id, head_revision_id, head_sequence, head_hash, schema_version]
      properties:
        novel_id:
          type: string
        head_revision_id:
          type: string
        head_sequence:
          type: integer
          minimum: 0
        head_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        schema_version:
          type: string
      additionalProperties: false
    CanonicalRevision:
      type: object
      required: [schema_version, novel_id, revision_id, parent_revision_id, chapters, created_at, content_hash]
      properties:
        schema_version:
          const: 1.0.0
        novel_id:
          type: string
        revision_id:
          type: string
        parent_revision_id:
          type: [string, "null"]
        chapters:
          type: array
          items:
            type: object
            additionalProperties: true
        created_at:
          type: string
          format: date-time
        content_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        extensions:
          type: object
          additionalProperties: true
      additionalProperties: false
...
    AccessKeyRequest:
      type: object
      required: [actor_id, scopes, expires_at]
      properties:
        actor_id:
          type: string
          pattern: '^[A-Za-z0-9._:@-]{1,128}$'
        scopes:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            enum: [novel:read, novel:analyze, novel:propose, novel:commit, novel:admin, style:analyze, style:admin, rewrite:propose, monitor:submit, worker:execute]
        expires_at:
          type: string
          format: date-time
      additionalProperties: false
    AccessKeyCreated:
      type: object
      required: [key_id, novel_id, actor_id, scopes, expires_at, secret]
      properties:
        key_id:
          type: string
          pattern: '^key_[0-9a-f-]{36}$'
        novel_id:
          type: string
          pattern: '^nov_[0-9a-f-]{36}$'
        actor_id:
          type: string
          pattern: '^[A-Za-z0-9._:@-]{1,128}$'
        scopes:
          type: array
          items:
            type: string
            enum: [novel:read, novel:analyze, novel:propose, novel:commit, novel:admin, style:analyze, style:admin, rewrite:propose, monitor:submit, worker:execute]
        expires_at:
          type: string
          format: date-time
        secret:
          type: string
          pattern: '^nv_key_[0-9a-f-]{36}\.[A-Za-z0-9_-]{43}$'
          readOnly: true
          description: One-time bearer credential; the response must not be cached.
      additionalProperties: false
    RevokeResponse:
      type: object
      required: [key_id, revoked]
      properties:
        key_id:
          type: string
          pattern: '^key_[0-9a-f-]{36}$'
        revoked:
          type: boolean
      additionalProperties: false
...